requirements.txt
appdirs==1.4.3
backports.weakref==1.0rc1
bleach==1.5.0
dlib==19.4.0
future==0.16.0
html5lib==0.9999999
Markdown==2.6.8
numpy==1.12.0
olefile==0.44
opencv-python==3.3.0
packaging==16.8
Pillow==4.0.0
protobuf==3.2.0
PyInstaller==3.2.1
pyparsing==2.2.0
pypiwin32==219
pyreadline==2.1
redis==2.10.6
six==1.10.0
tensorflow==1.2.1
tensorflow-tensorboard==0.1.8
Werkzeug==0.11.15
